refactor: move sitemap building
